APPLINK-14188 Fix Start and Stop sdl scripts
Start SDL should check that SDL started.
If it was not started, it should not create sdl.pid file

StopSDL should try stop SDL with sigterm, and wait 10 secconds.

+++ StartSDL.sh
@@ -1,5 +1,4 @@
 #!/bin/bash
-
 dirSDL=$1
 dirATF=$(pwd)
 appName=$2
@@ -7,8 +6,10 @@ cd $dirSDL
 LD_LIBRARY_PATH=$LD_LIBRARY_PATH:. export LD_LIBRARY_PATH
 ./$appName > /dev/null &
 sdl_pid=$!
+echo "SDL pid "$sdl_pid
+sleep 3
+test -e /proc/$sdl_pid || exit 1
 cd $dirATF
 echo $sdl_pid > sdl.pid
-sleep 3
 test -e sdl.pid && test -e /proc/$(cat sdl.pid) && exit 0
 exit 1
